+++ src/stories/Badge/Badge.stories.tsx
@@ -1,4 +1,4 @@
-import Badge from "../components/Badge/Badge";
+import Badge from "../../components/Badge/Badge";
 
 export default {
   component: Badge,
